+++ jobbers/adapters/redis/dead_queue.py
@@ -1,0 +1,191 @@
+"""
+Plain Redis dead-letter queue adapter.
+
+- `RedisDeadQueue` — DeadQueueProtocol + AtomicDeadQueueProtocol backed by plain Redis.
+"""
+
+from __future__ import annotations
+
+import datetime as dt
+from typing import TYPE_CHECKING, Any, cast
+
+from ulid import ULID
+
+from jobbers.models.task import Task
+
+if TYPE_CHECKING:
+    from collections.abc import Awaitable
+
+    from redis.asyncio.client import Pipeline, Redis
+
+    from jobbers.protocols import TaskStateProtocol, TransactionHandle
+
+
+class RedisDeadQueue:
+    r"""
+    DeadQueueProtocol and AtomicDeadQueueProtocol backed by plain Redis.
+
+    Reuses ``task:<task_id>`` keys for task data rather than duplicating blobs.
+
+    Keys:
+    - `dlq` sorted set — member: task_id bytes, score: failed_at Unix timestamp.
+    - `dlq-queue:{queue}` set — task_id bytes per queue name (queue filter index).
+    - `dlq-name:{name}` set — task_id bytes per task name (name filter index).
+    - `dlq-meta` hash — field: task_id bytes, value: `b"{queue}\0{name}\0{version}"`.
+    """
+
+    DLQ = "dlq"
+    DLQ_QUEUE = "dlq-queue:{queue}".format
+    DLQ_NAME = "dlq-name:{name}".format
+    DLQ_META = "dlq-meta"
+
+    def __init__(self, data_store: Redis, task_adapter: TaskStateProtocol) -> None:
+        self.data_store = data_store
+        self.ta = task_adapter
+
+    @property
+    def backend_key(self) -> str:
+        return str(id(self.data_store))
+
+    def pipeline(self, transaction: bool = True) -> Pipeline:
+        return self.data_store.pipeline(transaction=transaction)
+
+    async def ensure_index(self) -> None:
+        """No-op: plain-Redis dead queue does not use a search index."""
+
+    def stage_add(self, pipe: TransactionHandle, task: Task, failed_at: dt.datetime) -> None:
+        """Queue all four DLQ index writes onto pipe (no execute)."""
+        p: Any = pipe
+        p.zadd(self.DLQ, {bytes(task.id): failed_at.timestamp()})
+        p.sadd(self.DLQ_QUEUE(queue=task.queue), bytes(task.id))
+        p.sadd(self.DLQ_NAME(name=task.name), bytes(task.id))
+        p.hset(self.DLQ_META, str(task.id), f"{task.queue}\0{task.name}\0{task.version}")
+
+    async def add_to_dlq(self, task: Task, failed_at: dt.datetime) -> None:
+        """Add a task to the DLQ (non-pipeline version for saga path)."""
+        pipe = self.data_store.pipeline(transaction=True)
+        self.stage_add(pipe, task, failed_at)
+        await pipe.execute()
+
+    def stage_remove(self, pipe: TransactionHandle, task_id: ULID, queue: str, name: str) -> None:
+        """Queue all four DLQ index deletes onto pipe (no execute)."""
+        p: Any = pipe
+        p.zrem(self.DLQ, bytes(task_id))
+        p.srem(self.DLQ_QUEUE(queue=queue), bytes(task_id))
+        p.srem(self.DLQ_NAME(name=name), bytes(task_id))
+        p.hdel(self.DLQ_META, str(task_id))
+
+    async def remove_from_dlq(self, task_id: ULID, queue: str, name: str) -> None:
+        """Remove a task from the DLQ (non-pipeline version for saga path)."""
+        pipe = self.data_store.pipeline(transaction=True)
+        self.stage_remove(pipe, task_id, queue, name)
+        await pipe.execute()
+
+    async def get_history(self, task_id: str) -> list[dict[str, Any]]:
+        """Return the per-attempt error history for a DLQ task from its stored task blob."""
+        u = ULID.from_str(task_id)
+        task = await self.ta.get_task(u)
+        if task is None:
+            return []
+        return [{"attempt": i, "error": e} for i, e in enumerate(task.errors)]
+
+    async def get_by_ids(self, task_ids: list[str]) -> list[Task]:
+        """Fetch DLQ entries by explicit task ID list."""
+        if not task_ids:
+            return []
+        ulids = [ULID.from_str(tid) for tid in task_ids]
+        pipe = self.data_store.pipeline(transaction=False)
+        for u in ulids:
+            pipe.zscore(self.DLQ, bytes(u))
+        scores = await pipe.execute()
+        valid_ulids = [u for u, s in zip(ulids, scores) if s is not None]
+        if not valid_ulids:
+            return []
+        tasks: list[Task | None] = await self.ta.get_tasks_bulk(valid_ulids)
+        return [t for t in tasks if t is not None]
+
+    async def get_by_filter(
+        self,
+        queue: str | None = None,
+        task_name: str | None = None,
+        task_version: int | None = None,
+        limit: int = 100,
+    ) -> list[Task]:
+        """Fetch DLQ entries matching the given filter criteria."""
+        if queue is None and task_name is None:
+            id_bytes: list[bytes] = await cast(
+                "Awaitable[list[bytes]]",
+                self.data_store.zrevrange(self.DLQ, 0, -1),
+            )
+        else:
+            raw_ids: set[bytes]
+            if queue is not None and task_name is not None:
+                raw_ids = await cast(
+                    "Awaitable[set[bytes]]",
+                    self.data_store.sinter([self.DLQ_QUEUE(queue=queue), self.DLQ_NAME(name=task_name)]),
+                )
+            elif queue is not None:
+                raw_ids = await cast(
+                    "Awaitable[set[bytes]]",
+                    self.data_store.smembers(self.DLQ_QUEUE(queue=queue)),
+                )
+            else:
+                raw_ids = await cast(
+                    "Awaitable[set[bytes]]",
+                    self.data_store.smembers(self.DLQ_NAME(name=task_name)),
+                )
+            id_bytes = list(raw_ids)
+        if not id_bytes:
+            return []
+        ulid_list = [ULID.from_bytes(b) for b in id_bytes]
+        fetched: list[Task | None] = await self.ta.get_tasks_bulk(ulid_list)
+        results: list[Task] = []
+        for task in fetched:
+            if len(results) >= limit:
+                break
+            if task is None:
+                continue
+            if task_version is not None and task.version != task_version:
+                continue
+            results.append(task)
+        return results
+
+    async def remove_many(self, task_ids: list[str]) -> None:
+        """Remove multiple entries from the dead letter queue in a single transaction."""
+        if not task_ids:
+            return
+        ulids = [ULID.from_str(tid) for tid in task_ids]
+        read_pipe = self.data_store.pipeline(transaction=False)
+        for u in ulids:
+            read_pipe.hget(self.DLQ_META, str(u))
+        meta_list = await read_pipe.execute()
+        write_pipe = self.data_store.pipeline(transaction=True)
+        for u, meta_bytes in zip(ulids, meta_list):
+            if meta_bytes is None:
+                continue
+            queue, name, _ = meta_bytes.decode().split("\0")
+            self.stage_remove(write_pipe, u, queue, name)
+        await write_pipe.execute()
+
+    async def clean(self, earlier_than: dt.datetime) -> None:
+        """Remove failed tasks older than the specified datetime."""
+        old_ids: list[bytes] = await cast(
+            "Awaitable[list[bytes]]",
+            self.data_store.zrangebyscore(self.DLQ, "-inf", earlier_than.timestamp()),
+        )
+        if not old_ids:
+            return
+        old_ulid_strs = [str(ULID.from_bytes(b)) for b in old_ids]
+        pipe = self.data_store.pipeline(transaction=False)
+        for uid_str in old_ulid_strs:
+            pipe.hget(self.DLQ_META, uid_str)
+        meta_list = await pipe.execute()
+        pipe2 = self.data_store.pipeline(transaction=True)
+        for task_id_bytes, uid_str, meta_bytes in zip(old_ids, old_ulid_strs, meta_list):
+            pipe2.zrem(self.DLQ, task_id_bytes)
+            if meta_bytes is not None:
+                queue, name, _ = meta_bytes.decode().split("\0")
+                pipe2.srem(self.DLQ_QUEUE(queue=queue), task_id_bytes)
+                pipe2.srem(self.DLQ_NAME(name=name), task_id_bytes)
+                pipe2.hdel(self.DLQ_META, uid_str)
+        await pipe2.execute()
